src/test/java/ru/job4j/oop/TriangleTest.java
package ru.job4j.oop;

import static org.hamcrest.Matchers.closeTo;
import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertThat;

import org.hamcrest.MatcherAssert;
import org.junit.Assert;
import org.junit.Test;

public class TriangleTest {

    @Test
    public void when00and40and04Then8() {
        Point a = new Point(0, 0);
        Point b = new Point(4, 0);
        Point c = new Point(0, 4);
        Triangle triangle = new Triangle(a, b, c);
        double rsl = triangle.area();
        assertThat(rsl, closeTo(8, 0.001));
    }

    @Test
    public void when00and06and11Then6() {
        Point a = new Point(2, 2);
        Point b = new Point(4, 4);
        Point c = new Point(8, 2);
        Triangle triangle = new Triangle(a, b, c);
        double rsl = triangle.area();
        assertThat(rsl, closeTo(6, 0.001));
    }

    @Test
    public void when20and26and21ThenNotExist() {
        Point a = new Point(2, 0);
        Point b = new Point(2, 6);
        Point c = new Point(2, 1);
        Triangle triangle = new Triangle(a, b, c);
        double rsl = triangle.area();
        assertThat(rsl, closeTo(-1, 0.001));
    }

}